Remove gaps from entity file
