+++ pulse/editor/dual_volume_psd.py
@@ -39,7 +39,7 @@ def rotate_points(points, axis="along x-axis"):
     
     return rotated_points
 
-class DualChambersPSD:
+class DualVolumePSD:
     def __init__(self, device_data : dict) -> None:
 
         self.initialize()
@@ -50,6 +50,8 @@ class DualChambersPSD:
         self.pipe1_angle = None
         self.pipe2_angle = None
         self.segment_data = list()
+        self.acoustic_link_coords = dict()
+        self.structural_link_coords = dict()
 
     def unwrap_device_data(self, device_data : dict):
 
@@ -103,55 +105,35 @@ class DualChambersPSD:
 
     def get_axial_axial_segment_data(self):
 
+        inlet = np.array([0, 0, 0], dtype=float)
         versor_x = np.array([1, 0, 0], dtype=float)
         versor_z = np.array([0, 0, 1], dtype=float)
-        inlet = np.array([0, 0, 0], dtype=float)
+        offset_y = np.array([0, 0.01, 0], dtype=float)
 
-        if self.connection_pipe == "pipe #1":
+        P0 = inlet + versor_x * self.pipe1_length
+        Q0 = P0
+        Q1 = Q0 + versor_x * self.pipe3_distance
+        Q2 = Q0 + versor_x * self.volume1_length
+        Q3 = Q2 + versor_x * self.volumes_spacing
+        Q4 = Q1 + versor_x * self.pipe3_length
+        Q5 = Q3 + versor_x * self.volume2_length
+        P1 = Q0 + versor_x * self.pipe2_distance
+        outlet = Q5 + versor_x * self.pipe2_length
 
-            P0 = inlet + versor_x * self.pipe1_length
-            Q0 = P0
-            Q1 = Q0 + versor_x * self.pipe3_distance
-            Q2 = Q0 + versor_x * self.volume1_length
-            Q3 = Q2 + versor_x * self.volumes_spacing
-            Q4 = Q1 + versor_x * self.pipe3_length
-            Q5 = Q3 + versor_x * self.volume2_length
-            P1 = Q0 + versor_x * self.pipe2_distance
-            outlet = Q5 + versor_x * self.pipe2_length
-
-        else:
-
-            P0 = inlet - versor_x * self.pipe2_length
-            Q0 = P0
-            P1 = Q0 - versor_x * self.pipe1_distance
-            Q1 = Q0 - versor_x * self.pipe3_distance
-            Q2 = Q0 - versor_x * self.volume1_length
-            Q3 = Q2 - versor_x * self.volumes_spacing
-            Q4 = Q1 - versor_x * self.pipe3_length
-            Q5 = Q3 - versor_x * self.volume2_length
-            outlet = Q5 - versor_x * self.pipe1_length
-
-        base_points = [inlet, outlet, P0, P1, Q0, Q1, Q2, Q3, Q4, Q5]
+        base_points = np.array([inlet, outlet, P0, P1, Q0, Q1, Q2, Q3, Q4, Q5], dtype=float)
         rot_points = rotate_points(base_points, axis=self.axis)
         inlet, outlet, P0, P1, Q0, Q1, Q2, Q3, Q4, Q5 = translate_to_connection_point(rot_points, self.connection_point)
 
-        if self.connection_pipe == "pipe #1":
-            self.segment_data.append((inlet, Q0, self.pipe1_section_data))
-            self.segment_data.append((outlet, Q1, self.pipe2_section_data))
-            self.segment_data.append((Q0, Q2, self.volume1_section_data))
-            self.segment_data.append((Q3, Q5, self.volume2_section_data))
-            self.segment_data.append((Q1, Q2, self.pipe3_section_data))
-            self.segment_data.append((Q2, Q3, self.pipe3_section_data))
-            self.segment_data.append((Q3, Q4, self.pipe3_section_data))
+        self.acoustic_link_coords[1] = Q1
+        self.acoustic_link_coords[2] = Q4
 
-        else:
-            self.segment_data.append((inlet, Q0, self.pipe2_section_data))
-            self.segment_data.append((outlet, Q1, self.pipe1_section_data))
-            self.segment_data.append((Q0, Q2, self.volume2_section_data))
-            self.segment_data.append((Q3, Q5, self.volume1_section_data))
-            self.segment_data.append((Q1, Q2, self.pipe3_section_data))
-            self.segment_data.append((Q2, Q3, self.pipe3_section_data))
-            self.segment_data.append((Q3, Q4, self.pipe3_section_data))
+        self.segment_data.append((inlet, Q0, self.pipe1_section_data))
+        self.segment_data.append((outlet, Q1, self.pipe2_section_data))
+        self.segment_data.append((Q0, Q2, self.volume1_section_data))
+        self.segment_data.append((Q3, Q5, self.volume2_section_data))
+        self.segment_data.append((Q1, Q2, self.pipe3_section_data))
+        self.segment_data.append((Q2, Q3, self.pipe3_section_data))
+        self.segment_data.append((Q3, Q4, self.pipe3_section_data))
 
     def get_axial_radial_segment_data(self):
 
@@ -189,9 +171,10 @@ class DualChambersPSD:
 
     def get_radial_radial_segment_data(self):
 
+        inlet = np.array([0, 0, 0], dtype=float)
         versor_x = np.array([1, 0, 0], dtype=float)
         versor_z = np.array([0, 0, 1], dtype=float)
-        inlet = np.array([0, 0, 0], dtype=float)
+        offset_y = np.array([0, 0.01, 0], dtype=float)
 
         pipe1_angle = self.pipe1_angle * (np.pi / 180)
         rot_pipe1 =  rotation_about_x_axis(pipe1_angle) @ versor_z
@@ -199,55 +182,46 @@ class DualChambersPSD:
         pipe2_angle = self.pipe2_angle * (np.pi / 180)
         rot_pipe2 =  rotation_about_x_axis(pipe2_angle) @ versor_z
 
-        if self.connection_pipe == "pipe #1":
+        P0 = inlet - rot_pipe1 * self.pipe1_length
+        Q0 = P0 - versor_x * self.pipe1_distance
+        P1 = Q0 + versor_x * self.pipe2_distance
+        Q1 = Q0 + versor_x * self.pipe3_distance
+        Q2 = Q0 + versor_x * self.volume1_length
+        Q3 = Q2 + versor_x * self.volumes_spacing
+        Q4 = Q1 + versor_x * self.pipe3_length
+        Q5 = Q3 + versor_x * self.volume2_length
+        outlet = P1 + rot_pipe2 * self.pipe2_length
 
-            P0 = inlet - rot_pipe1 * self.pipe1_length
-            Q0 = P0 - versor_x * self.pipe1_distance
-            P1 = Q0 + versor_x * self.pipe2_distance
-            Q1 = Q0 + versor_x * self.pipe3_distance
-            Q2 = Q0 + versor_x * self.volume1_length
-            Q3 = Q2 + versor_x * self.volumes_spacing
-            Q4 = Q1 + versor_x * self.pipe3_length
-            Q5 = Q3 + versor_x * self.volume2_length
-            outlet = P1 + rot_pipe2 * self.pipe2_length
+        Q1o = Q1 + offset_y
+        Q2o = Q2 + offset_y
+        Q3o = Q3 + offset_y
+        Q4o = Q4 + offset_y
 
-        else:
-
-            P0 = inlet - rot_pipe2 * self.pipe2_length
-            Q0 = P0 - versor_x * self.pipe2_distance
-            P1 = Q0 + versor_x * self.pipe1_distance
-            Q1 = Q0 + versor_x * self.pipe3_distance
-            Q2 = Q0 + versor_x * self.volume1_length
-            Q3 = Q2 + versor_x * self.volumes_spacing
-            Q4 = Q1 + versor_x * self.pipe3_length
-            Q5 = Q3 + versor_x * self.volume2_length
-            outlet = P1 + rot_pipe1 * self.pipe2_length            
-
-        base_points = [inlet, outlet, P0, P1, Q0, Q1, Q2, Q3, Q4, Q5]
+        base_points = np.array([inlet, outlet, P0, P1, Q0, Q1, Q2, Q3, Q4, Q5, Q1o, Q2o, Q3o, Q4o], dtype=float)
         rot_points = rotate_points(base_points, axis=self.axis)
-        inlet, outlet, P0, P1, Q0, Q1, Q2, Q3, Q4, Q5 = translate_to_connection_point(rot_points, self.connection_point)
+        inlet, outlet, P0, P1, Q0, Q1, Q2, Q3, Q4, Q5, Q1o, Q2o, Q3o, Q4o = translate_to_connection_point(rot_points, self.connection_point)
 
-        if self.connection_pipe == "pipe #1":
-            self.segment_data.append((inlet, P0, self.pipe1_section_data))
-            self.segment_data.append((outlet, P1, self.pipe2_section_data))
-            self.segment_data.append((Q0, P0, self.volume1_section_data))
-            self.segment_data.append((P0, Q2, self.volume1_section_data))
-            self.segment_data.append((Q3, P1, self.volume2_section_data))
-            self.segment_data.append((P1, Q5, self.volume2_section_data))
-            self.segment_data.append((Q1, Q2, self.pipe3_section_data))
-            self.segment_data.append((Q2, Q3, self.pipe3_section_data))
-            self.segment_data.append((Q3, Q4, self.pipe3_section_data))
+        self.acoustic_link_coords[1] = (Q1, Q1o)
+        self.acoustic_link_coords[2] = (Q4, Q4o)
 
-        else:
-            self.segment_data.append((inlet, P0, self.pipe1_section_data))
-            self.segment_data.append((outlet, P1, self.pipe2_section_data))
-            self.segment_data.append((Q0, P1, self.volume1_section_data))
-            self.segment_data.append((P1, Q2, self.volume1_section_data))
-            self.segment_data.append((Q3, P0, self.volume2_section_data))
-            self.segment_data.append((P0, Q5, self.volume2_section_data))
-            self.segment_data.append((Q1, Q2, self.pipe3_section_data))
-            self.segment_data.append((Q2, Q3, self.pipe3_section_data))
-            self.segment_data.append((Q3, Q4, self.pipe3_section_data))
+        self.structural_link_coords[1] = (Q2, Q2o)
+        self.structural_link_coords[1] = (Q3, Q3o)
+
+        self.segment_data.append((inlet, P0, self.pipe1_section_data))
+        self.segment_data.append((outlet, P1, self.pipe2_section_data))
+        self.segment_data.append((Q0, P0, self.volume1_section_data))
+        self.segment_data.append((P0, Q1, self.volume1_section_data))
+        self.segment_data.append((Q1, Q2, self.volume1_section_data))
+        self.segment_data.append((Q3, Q4, self.volume2_section_data))
+        self.segment_data.append((Q4, P1, self.volume2_section_data))
+        self.segment_data.append((P1, Q5, self.volume2_section_data))
+        self.segment_data.append((Q1o, Q2o, self.pipe3_section_data))
+        self.segment_data.append((Q2o, Q3o, self.pipe3_section_data))
+        self.segment_data.append((Q3o, Q4o, self.pipe3_section_data))
+        self.segment_data.append((Q1, Q1o, "acoustic_link"))
+        self.segment_data.append((Q4, Q4o, "acoustic_link"))
+        self.segment_data.append((Q2, Q2o, "structural_link"))
+        self.segment_data.append((Q3, Q3o, "structural_link"))
 
     def process_segment_data(self):
