+++ pulse/interface/handler/geometry_handler.py
@@ -215,6 +215,9 @@ class GeometryHandler:
                     pipe.set_diameter(initial_diameter, final_diameter)
 
                 structures.append(pipe)
+            
+                if "psd_label" in data.keys():
+                    pipe.extra_info["psd_label"] = data["psd_label"]
 
         pipeline = app().geometry_toolbox.pipeline
         pipeline.structures.clear()
@@ -532,7 +535,6 @@ class GeometryHandler:
 
         PrintMessageInput([window_title_2, title, message])
 
-
     def export_entity_file(self):
 
         tag = 1
@@ -540,6 +542,7 @@ class GeometryHandler:
         section_info = dict()
         element_type_info = dict()
         material_info = dict()
+        psd_info = dict()
         pipeline = app().geometry_toolbox.pipeline
 
         for structure in pipeline.structures:
@@ -564,6 +567,9 @@ class GeometryHandler:
                 if structure.extra_info["structural_element_type"] is not None:
                     element_type_info[tag] = structure.extra_info["structural_element_type"]
 
+            if "psd_label" in structure.extra_info.keys():
+                psd_info[tag] = structure.extra_info["psd_label"]
+
             tag += 1
 
         if len(points_info):
@@ -587,6 +593,10 @@ class GeometryHandler:
             if len(material_info):
                 for tag, material_id in material_info.items():
                     self.file.add_material_segment_in_file(tag, material_id)
+
+            if psd_info:
+                for tag, label in psd_info.items():
+                    self.file.add_psd_label_in_file(tag, label)
 
             self.file.modify_project_attributes(import_type = 1)
             # self.load_project()
